+++ dev/docker-mariadb/init.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+echo "Waiting for redis and mariadb to start up..."
+bash ./dev/docker-common/wait.sh redis:6379
+bash ./dev/docker-common/wait.sh db:3306
+echo "=== READY ==="
+tail -f /dev/null