+++ server/models/renderers.js
@@ -22,10 +22,13 @@ module.exports = class Renderer extends Model {
 
       properties: {
         key: {type: 'string'},
-        isEnabled: {type: 'boolean'},
-        config: {type: 'object'}
+        isEnabled: {type: 'boolean'}
       }
     }
+  }
+
+  static get jsonAttributes() {
+    return ['config']
   }
 
   static async getRenderers() {